refactor: streamline creator redemption admin flows
