+++ docs/exec-plans/completed/admin-orders-page-slimming.md
