+++ docs/exec-plans/completed/operator-user-detail-page-slimming.md
