+++ src/api/flaskr/service/promo/api.py
@@ -1,15 +1,10 @@
 from __future__ import annotations
 
 from flaskr.service.promo.admin import (
-    create_creator_course_redemption_coupon,
     create_operator_promotion_campaign,
     create_operator_promotion_coupon,
-    get_creator_course_redemption_coupon_detail,
     get_operator_promotion_campaign_detail,
     get_operator_promotion_coupon_detail,
-    list_creator_course_redemption_coupons,
-    list_creator_course_redemption_coupon_codes,
-    list_creator_course_redemption_coupon_usages,
     list_operator_promotion_campaign_redemptions,
     list_operator_promotion_campaigns,
     list_operator_promotion_coupon_codes,
@@ -17,10 +12,17 @@ from flaskr.service.promo.admin import (
     list_operator_promotion_coupons,
     update_operator_promotion_campaign,
     update_operator_promotion_campaign_status,
-    update_creator_course_redemption_coupon,
-    update_creator_course_redemption_coupon_status,
     update_operator_promotion_coupon,
     update_operator_promotion_coupon_status,
+)
+from flaskr.service.promo.creator_redemption import (
+    create_creator_course_redemption_coupon,
+    get_creator_course_redemption_coupon_detail,
+    list_creator_course_redemption_coupons,
+    list_creator_course_redemption_coupon_codes,
+    list_creator_course_redemption_coupon_usages,
+    update_creator_course_redemption_coupon,
+    update_creator_course_redemption_coupon_status,
 )
 from flaskr.service.promo.funcs import (
     build_campaign_enabled_expression,
